requirements.txt
sounddevice
soundfile
openai-whisper
mcp[cli]
fastmcp